Implemented a functional window into engine
The engine now has a window class that has the ability to be platform agnostic depending on the api. currently i have of course only implemented windows, and will test further once i install linux onto one of my seperate pc's.

i also decided to add application to the engine as well so that i can keep many things on the side of the engine, while still allowing gameside to have influence in the form of creating a public app child.