created an application paradigm, and debugged a low frame rate issue
basically now app is part of the engine, and it is up to the game to instantiate a child class of the app. this means that the app can take care of all engine side programs/systems and the game can just focus on being the game. also i found a low framerate problem and debugged it, it was related to swapbuffers and not yet having a rendering api in place to actually swap.